moved dnd context to separate component
