+++ package.json
@@ -23,7 +23,8 @@
 		"react-redux": "^8.0.5",
 		"react-select": "^5.7.0",
 		"styled-components": "^5.3.6",
-		"styled-system": "^5.1.5"
+		"styled-system": "^5.1.5",
+		"use-react-screenshot": "^3.0.0"
 	},
 	"devDependencies": {
 		"@types/chroma-js": "^2.1.4",
